+++ src/internal/document_api/mod.rs
@@ -46,7 +46,7 @@ const HEADER_META_LENGTH_LENGTH: usize = 2;
 const CURRENT_DOCUMENT_ID_VERSION: u8 = 2;
 
 /// Document ID. Unique within the segment. Must match the regex `^[a-zA-Z0-9_.$#|@/:;=+'-]+$`
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct DocumentId(pub(crate) String);
 impl DocumentId {
     pub fn id(&self) -> &str {
@@ -74,7 +74,7 @@ impl TryFrom<String> for DocumentId {
 }
 
 /// (unencrypted) name of a document. Construct via `try_from(&str)`
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct DocumentName(pub(crate) String);
 impl DocumentName {
     pub fn name(&self) -> &String {
@@ -92,7 +92,7 @@ impl TryFrom<&str> for DocumentName {
 struct DocHeaderPacked(Vec<u8>);
 
 /// Represents a parsed document header which is decoded from JSON
-#[derive(Debug, Serialize, Deserialize, PartialEq)]
+#[derive(Debug, PartialEq, Serialize, Deserialize)]
 struct DocumentHeader {
     #[serde(rename = "_did_")]
     document_id: DocumentId,
@@ -161,7 +161,7 @@ fn parse_document_parts(
 }
 
 /// Represents the reason a document can be viewed by the requesting user.
-#[derive(Serialize, Deserialize, Debug, Clone, Hash, PartialEq, Eq)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub enum AssociationType {
     /// User created the document
@@ -173,7 +173,7 @@ pub enum AssociationType {
 }
 
 /// Represents a User struct which is returned from doc get to show the IDs of users the document is visible to
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct VisibleUser {
     id: UserId,
 }
@@ -184,7 +184,7 @@ impl VisibleUser {
 }
 
 /// Represents a Group struct which is returned from doc get to show the IDs and names of groups the document is visible to
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub struct VisibleGroup {
     id: GroupId,
     name: Option<GroupName>,
@@ -201,7 +201,7 @@ impl VisibleGroup {
 /// Single document's (abbreviated) metadata. Returned as part of a `DocumentListResult`.
 ///
 /// If you want full metadata for a document, see `DocumentMetadataResult`
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct DocumentListMeta(DocumentListApiResponseItem);
 impl DocumentListMeta {
     pub fn id(&self) -> &DocumentId {
@@ -222,7 +222,7 @@ impl DocumentListMeta {
 }
 
 /// Metadata for each of the documents that the current user has access to decrypt.
-#[derive(Debug)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct DocumentListResult {
     result: Vec<DocumentListMeta>,
 }
@@ -233,7 +233,7 @@ impl DocumentListResult {
 }
 
 /// Full metadata for a document.
-#[derive(Clone)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct DocumentMetadataResult(DocumentMetaApiResponse);
 impl DocumentMetadataResult {
     pub fn id(&self) -> &DocumentId {
@@ -273,7 +273,7 @@ impl DocumentMetadataResult {
 /// - `encrypted_data` - Bytes of encrypted document content
 /// - `encrypted_deks` - List of encrypted document encryption keys (EDEK) of users/groups that have been granted access to `encrypted_data`
 /// - `access_errs` - Users and groups that could not be granted access
-#[derive(Debug)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct DocumentEncryptUnmanagedResult {
     id: DocumentId,
     encrypted_data: Vec<u8>,
@@ -329,7 +329,7 @@ impl DocumentEncryptUnmanagedResult {
 /// - `encrypted_data` - Bytes of encrypted document content
 /// - `grants` - Users and groups that have access to decrypt the `encrypted_data`
 /// - `access_errs` - Users and groups that could not be granted access
-#[derive(Debug)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct DocumentEncryptResult {
     id: DocumentId,
     name: Option<DocumentName>,
@@ -363,7 +363,7 @@ impl DocumentEncryptResult {
     }
 }
 /// Result of decrypting a document. Includes minimal metadata as well as the decrypted bytes.
-#[derive(Debug, PartialEq, Clone)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct DocumentDecryptResult {
     id: DocumentId,
     name: Option<DocumentName>,
@@ -390,7 +390,7 @@ impl DocumentDecryptResult {
 }
 
 /// A failure to edit the access list of a document.
-#[derive(Debug, Clone)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct DocAccessEditErr {
     /// User or group whose access was to be granted/revoked
     pub user_or_group: UserOrGroup,
@@ -409,7 +409,7 @@ impl DocAccessEditErr {
 
 /// Result of granting or revoking access to a document. Both grant and revoke support partial
 /// success.
-#[derive(Debug)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct DocumentAccessResult {
     succeeded: Vec<UserOrGroup>,
     failed: Vec<DocAccessEditErr>,
@@ -433,11 +433,11 @@ impl DocumentAccessResult {
         &self.failed
     }
 }
-#[derive(Clone, PartialEq, Debug)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 struct DecryptedData(Vec<u8>);
 
 /// Result of successful unmanaged decryption
-#[derive(Clone, PartialEq, Debug)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct DocumentDecryptUnmanagedResult {
     id: DocumentId,
     access_via: UserOrGroup,
@@ -462,7 +462,7 @@ impl DocumentDecryptUnmanagedResult {
 }
 
 /// Either a user or a group. Allows for containing both.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase", tag = "type")]
 pub enum UserOrGroup {
     User { id: UserId },
@@ -653,17 +653,26 @@ where
         Ok((vec![], cached_policy.clone()))
     } else {
         // otherwise query the webservice and cache the result if there are no errors
-        get_policy_f.await.map(|policy_resp| {
-            let (errs, public_keys) = process_policy(&policy_resp);
-            if errs.is_empty() {
-                //if the cache has grown too large, clear it prior to adding new entries
-                if policy_cache.len() >= config.max_entries {
-                    policy_cache.clear()
+        get_policy_f
+            .await
+            .map(|policy_resp| {
+                let (errs, public_keys) = process_policy(&policy_resp);
+                if errs.is_empty() {
+                    //if the cache has grown too large, clear it prior to adding new entries
+                    if policy_cache.len() >= config.max_entries {
+                        policy_cache.clear()
+                    }
+                    policy_cache.insert(grant.clone(), public_keys.clone());
                 }
-                policy_cache.insert(grant.clone(), public_keys.clone());
-            }
-            (errs, public_keys)
-        })
+                (errs, public_keys)
+            })
+            .map_err(|x| match x {
+                IronOxideErr::RequestError {
+                    http_status: Some(code),
+                    ..
+                } if code == 404 => IronOxideErr::PolicyDoesNotExist,
+                e => e,
+            })
     }
 }
 
@@ -785,7 +794,7 @@ fn recrypt_document<CR: rand::CryptoRng + rand::RngCore>(
 /// Once decrypted, the DEK serves as a symmetric encryption key.
 ///
 /// It can also be useful to think of an EDEK as representing a "document access grant" to a user/group.
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct EncryptedDek {
     grant_to: WithKey<UserOrGroup>,
     encrypted_dek_data: recrypt::api::EncryptedValue,
@@ -859,7 +868,7 @@ impl TryFrom<&EncryptedDek> for EncryptedDekP {
 /// Result of recrypt encryption. Contains the encrypted DEKs and the encrypted (user) data.
 /// `RecryptionResult` is an intermediate value as it cannot be serialized to bytes directly.
 /// To serialize to bytes, first construct an `EncryptedDoc`
-#[derive(Debug, Clone)]
+#[derive(Clone, Debug)]
 struct RecryptionResult {
     edeks: Vec<EncryptedDek>,
     encrypted_data: AesEncryptedValue,
@@ -1287,6 +1296,7 @@ mod tests {
     use galvanic_assert::matchers::{collection::*, *};
 
     use super::*;
+    use crate::internal::RequestErrorCode;
     use dashmap::DashMap;
     use std::borrow::Borrow;
 
@@ -1324,6 +1334,30 @@ mod tests {
         })
         .await?;
         assert_eq!(1, policy_cache.len());
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn policy_404_gives_nice_error() -> Result<(), IronOxideErr> {
+        let policy_grant = PolicyGrant::default();
+        let policy_cache = DashMap::new();
+        let config = PolicyCachingConfig::default();
+
+        // show transformation of RequestError - 404 for Policy GET to PolicyDoesNotExist
+        let err_result = get_cached_policy_or(&config, &policy_grant, &policy_cache, async {
+            Err(IronOxideErr::RequestError {
+                message: "".into(),
+                code: RequestErrorCode::PolicyGet,
+                http_status: Some(404),
+            })
+        })
+        .await;
+        assert!(err_result.is_err());
+        assert_that!(
+            &err_result.unwrap_err(),
+            is_variant!(IronOxideErr::PolicyDoesNotExist)
+        );
 
         Ok(())
     }